src: Add back minor v4 changes for hotfix.
- https://github.com/samuel-lucas6/Kryptor/commit/57bc3ea089e528cc558e4f7a59b1dcc15c8582d7
- https://github.com/samuel-lucas6/Kryptor/commit/d8c2986148674b0ed01319fcdb8fcbdd729f20ab
- https://github.com/samuel-lucas6/Kryptor/commit/7d273c54a2ed9f26398ae476898487aba4e7f66c
- https://github.com/samuel-lucas6/Kryptor/commit/a4d40ad9fcb8a7aa5a8d5732535228680cfed7a9
- https://github.com/samuel-lucas6/Kryptor/commit/84ab7f6fe7ee62450fa145d63118328af79bcfa3
- https://github.com/samuel-lucas6/Kryptor/commit/6c6b04aaf82ecb0debb81b33627e72e0f9dbd299
- https://github.com/samuel-lucas6/Kryptor/commit/d3282f0389635a045af0e79ab20b7c80e2ea08ab
- https://github.com/samuel-lucas6/Kryptor/commit/0aedb7fe620b228a010a9cde1622db2d51ee119a

+++ src/KryptorCLI/FileEncryption/DecryptFile.cs
@@ -47,7 +47,7 @@ public static class DecryptFile
             inputFile.Dispose();
             Globals.SuccessfulCount += 1;
             RestoreFileName.RenameFile(outputFilePath, fileNameLength);
-            FileHandling.DeleteFile(inputFile.Name);
+            if (Globals.Overwrite) { FileHandling.DeleteFile(inputFile.Name); }
         }
         catch (Exception ex) when (ExceptionFilters.Cryptography(ex))
         {
